Prevent deadlock during `InflightRequestes->cancelAll()`
Canceling requests will invoke `Request::callback()` that will execute
`Request::setStatus` and attempt to remove it from `InflightRequests`
being a cause for deadlock. By adding a new cancelation mutex and
checking if it can be locked we may eliminate both calls attempting to
acquire `_mutex` from the same thread and cause a deadlock.

Another idea was to use a single reentrant lock `InflightRequests`, but
that would be problematic as it would still permit inserting and
removing objects simultaneously, thus causign other issues, therefore
that idea was discarded.