Fix cancelation of requests during worker progress
Before canceling any requests it is necessary to progress the worker.
This is required to ensure incoming messages are received before they
are canceled. For example, during the endpoint error callback, inflight
requests must be registered for cancelation and if the worker is not
progressed but the remote endpoint successfully delivered messages,
those messages will then be canceled and lost.