Unify conda recipes into a single split recipe
With this change it's now possible to generate standalone `libucxx`
conda package containing only C++ library and header files. On top of
that `libucxx-examples` and `libucxx-tests` are built as well, being
also C++-only executables. Finally, the `ucxx` package contains the
Python component of the C++ library, together with Cython bindings and
the Python libraries.