Signal the worker progress event from the worker progress thread
Without signaling the worker progress event from the worker progress
thread while shutting down the thread it may occur that the thread hangs
as the event never gets awaken due to the lack of tasks to progress.
This change ensures the worker progress event is signaled before joining
the thread.