Avoid null pointer dereferencing in user callback request tests
When a request completes immediately, the callback is executed before
the request method returns, and thus it isn't possible to get the status
of the object for which we don't have a reference yet. For now we just
hack around those cases (which happen infrequently), but in the future
we should find a better way to test this.