Add a timeout argument to `progressWorkerEvent` (#30)
Instead of blocking indefinitely until an evens occurs, allows the user to specify a timeout the epoll wait operation, causing it to return.

In some cases using blocking mode in the progress thread may cause the event to never wake, allowing it to break periodically fixes hangs in those cases. It's not clear whether this is an abuse of blocking progress mode in UCX, but for now this resolves that issue.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/ucxx/pull/30