Pass request's status directly onto user-defined callback
To ensure that a user-defined request callback has access to the status,
an argument containing the status is now added to the callback function
prototype. This allows the callback to know the request status before
the `ucxx::Request`s is set, and thus preventing the application from
believing the request has completed while the callback did not run yet.

To make this change more evident and organized, new types
`RequestCallbackUserFunction` and `RequestCallbackUserData` are added.