Move `ucxx::Request` status setting to base class
Setting the status from a derived class before `callback()`, and,
therefore `setStatus()`, can be dangerous in situations where the user
code is not holding a reference to the request. In such cases, the user
code may be checking for the status to change to immediately continue by
releasing the request, when that happens and `setStatus()` is eventually
called, that will be operating on an invalid `ucxx::Request` object.