Use the status of first failing request in `ucxx::RequestTagMulti`
The status of `ucxx::RequestTagMulti` was until now always
`UCS_INPROGRESS` or `UCS_OK` after completing. This is not right but
there is no good way to combine the statuses of all underlying
requests. Therefore we now set the status of the first failing request
as the final status instead of `UCS_OK` if at least one of the requests
failed. The user can still check each underlying's request status if
granular information is required.