Execute user callback within `ucxx::Request::setStatus()`
By executing the user-defined request callback within the scope of
`ucxx::Request::setStatus()` we prevent accidental reordering mistakes,
as the callback needs to always execute after the status is set, as the
callback itself might make use of that information. Additionally it
will prevent missing execution of the callback should `setStatus()` be
called elsewhere in the future.