Introduce mutex in `ucxx::Request`
When running with the progress thread, concurrency issues may happen,
which a mutex can resolve in `ucxx::Request`. A common case is when
`ucxx::Endpoint` is registering an inflight request and `setStatus()`
has just begun running, in which case the inflight request will be
removed (a no-op as it hasn't been previously registered) and then
actually registered, which will cause the `ucxx::Request` to never
cleanup.