Prevent early destruction of `ucxx::Endpoint`
When the last request to hold `std::shared_ptr<ucxx::Endpoint>` is
erased it may cause the `ucxx::Endpoint`s destructor and subsequently
the `close()` method to be called which will in turn call `cancelAll()`
and attempt to take the mutexes. For this reason a temporary copy of the
request being erased should be held to allow unlocking the mutexes and
only then destroy the object upon `ucxx::InflightRequests::remove()`'s
return.