Timeout generic callbacks used in destructors and inflight request cancelation (#123)
Destructors may be called from Python's garbage collector, disallowing us to disable the GIL for that operation. When that occurs, there might be a deadlock as other threads have the GIL, such as the progress thread when a Python's listener callback is invoked and executed in that thread.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/ucxx/pull/123