Clear out callback data once we have run the callback
Although the UCXX python layer is careful not to set up
refcycles (such that everything is collected deterministically), the
callbacks are a loophole that can break this guarantee.

It is easy to set up a callback that references an object that
produces a reference cycle. To avoid this, once we are done,
internally, with the callback function (after it is called) drop any
references to it.

This reduces the ease of post-hoc debugging (since one the callback
has run, we can't see what it was), but ensures that we don't
accidentally have a refcycle that inhibits deterministic collection of
UCXX objects.