Simplify `ucxx::RequestTagMulti` final status logic
The current way to handle the final status of `ucxx::RequestTagMulti`
involves looping all requests, this change simplifies that by simply
caching the status of the first request error. Additionally this
prevents checking the status of the final request if it hasn't been set
yet, which would then return `UCS_INPROGRESS` instead.