Use a weakref finalizer to stop notifier thread
When creating our own application context with the python notifier
thread enabled we previously had to arrange to stop it before dropping
the reference to the context we created. The consequence of not doing
so is that our application hangs in exit waiting for a thread to
join that never will.

Applying the principle of least surprise, an object should clean up
all of its resources when dropping out of scope. To ensure this
change Application.stop_notifier_thread to a staticmethod and attach
it as a weakref finalizer when starting the notifier thread. This way,
we don't have to handle the case where there is no notifier thread,
nor do we have to remember to call the stop function manually.

This additionally allows us to clean up the top-level reset function
since there is no longer a need to stop the notifier "manually" during
reset.