Prevent pointer conversion within `@classmethod`
Converting `shared_ptr`->`void*`->`uintptr_t` is a very cumbersome and
error prone solution. By instantiating the object within the
`@classmethod` scope we can thus prevent calling the constructor again
where the conversion is necessary. The downside is we now must set
memeber attributes also in the scope of `@classmethod`, but that should
be a better tradeoff to the pointer conversion.