Unset Python Endpoint close callback upon destruction
In some cases, particularly with `distributed-ucxx`, the C++ Endpoint
outlives the Python object that owns a close callback registered with
`set_close_callback()`, making it an invalid reference and causing
segfaults. To prevent that it is necessary to remove the close callback,
preventing it from being called after it's not valid anymore.