Unset Python Endpoint close callback upon destruction (#206)
In some cases, particularly with `distributed-ucxx`, the C++ Endpoint outlives the Python object that owns a close callback registered with `set_close_callback()`, making it an invalid reference and causing segfaults. To prevent that it is necessary to remove the close callback, preventing it from being called after it's not valid anymore.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/ucxx/pull/206