Use `weak_ptr<ucxx::Endpoint>` in `ErrorCallbackData`
A race could cause the `Endpoint::errorCallback` to dereference an
already closed and destroyed `ucxx::Endpoint`, leading to segmentation
faults and sometimes deadlocks. With the use of a `weak_ptr<Endpoint>`
that can be prevented by returning before dereferencing the pointer
which should be safe since the owner has already been destroyed anyway.