Use simple class instance of `ucxx::WorkerProgressThread`
Instantiate `ucxx::WorkerProgressThread` without a smart pointer. With
that we can have more control of whether the thread has stopped.
Additionally move away from a reference for the stop flag that seems to
be overwritten or not properly passed to the thread and thus making its
content unreliable, a smart pointer seems to work better for that case.