Add several fixes primarily targeting worker progress thread (#238)
# Add several fixes primarily targeting worker progress thread

Several changes are added to improve stability of worker progress thread. Details of changes are below.

## Support generic callback cancelation

Once registered it was not possible to cancel generic callbacks, which could cause the callback to be called after the timeout had expired and the owner/resources were not alive anymore. This change allows canceling a generic callback that couldn't be completed successfully, avoiding access to invalid memory of objects that are not alive anymore.

## Remove `ErrorCallbackData`

The `ErrorCallbackData` structure and `ucxx::Endpoint::errorCallback` static method used by `ucxx::Endpoint` were removed in favor of using a friend function `endpointErrorCallback` that receives a pointer to `ucxx::Endpoint` itself, thus simplifying ownership of all resources where only `ucxx::Endpoint` now is responsible to retain ownership of all resources.

## Improvements to `ucxx::InflightRequests`

### Add missing locks

All accesses to any of internal `_trackedRequests` resources are now protected by locks. This has been the source of various issues in the past, thus switching to a safer approach seems wise.

During `merge()` both merger and mergee have resource locks now to prevent races.

### Use regular instances instead of pointer for `TrackedRequests`

To simplify `ucxx::InflightRequests`/`TrackedRequests` instead of using pointers to `InflightRequestsMap` objects, switch to using simple instances.

## Improvements to `ucxx::WorkerProgressThread`

### Replace `std::ref` with `std::shared_ptr`

Using `std::ref` was causing undefined behavior in the state of `_stop`, switching to `std::shared_ptr` has resolved the problem.

### Set progress thread ID before progress loop stats

Set the thread ID with parent before starting the progress loop, so that checking whether the a callback is running in the progress thread is always correct, otherwise the parent may be unable to determine the progress thread has already started.

### Use simple instance instead of pointer

New members, such as `isRunning()` and `stop()`, were introduced so a simple instance of `WorkerProgressThread` may be used instead of smart pointers. With this change the owner of the instance may query whether the thread indeed stopped, as opposed to just letting it go out-of-scope while the thread was not joined.

## Improvement to `CallbackNotifier`

Instead of blocking continuously for longer periods of time, it's now possible to periodically signal the worker again to cause it to wakeup (when in blocking mode). Currently defaults to 100ms.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/ucxx/pull/238