add telemetry (#321)
Adds telemetry that will send timing information about each step in the build process to a time series database, for later viewing and analytics.

Tracking issue: https://github.com/rapidsai/build-infra/issues/139

This is not expected to work right now. We are in flux regarding a time series database that is in the same private network as the build machines. @ajschmidt8 has restricted the mTLS certificates that are used for the current database to specific repos. Either he will enable those here, or we'll make the switch to the new server, and then this mechanism should work.

Authors:
  - Mike Sarahan (https://github.com/msarahan)
  - James Lamb (https://github.com/jameslamb)

Approvers:
  - James Lamb (https://github.com/jameslamb)

URL: https://github.com/rapidsai/ucxx/pull/321