Expose class structure in `libucxx.pxd` (#372)
Expose class structure in `libucxx.pxd`. This change is useful so that projects building on top of the Python interface can take `shared_ptr<ucxx::Component>` objects and plug those objects into their C++ backends.

Although it's not possible to define private attributes in a `cdef class`, we attempt to mark them private with the leading underscore in their names, such as `UCXContext._context`, and then provide a `UCXContext.get_ucxx_shared_ptr` for those members that are expected to be externally usable.

Testing will be handled as part of #376.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/ucxx/pull/372