Add retry mechanism to `test_ucx_config_w_env_var` (#377)
Add a retry mechanism to ensure the Dask scheduler will bind to a free port and retry if the port is not free. We still use `distributed.utils.open_port` to attempt first, but that port seems not to be realiable for UCX, in which cases `UCXXBusyError` is raised, we capture that, increment the port and retry.

This test fails seldom in CI for the reason stated above.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/ucxx/pull/377