Use `string_view` in address methods
When setting or getting the underlying string data used for an `Address`
object, avoid extra copying by using `string_view`. During
initialization a copy is performed into the `Address` object's local
stroage, which makes sense. Though the input to the function need not
copy to `string`. It can just be viewed while the data is copied into
the object's local memory. Further when converting the `Address` to a
string, a view can be used to avoid a copy during the return. If the
caller would like to persist the string longer than the lifetime of the
`Address` object, they can copy it at that point. Though we can save
them a copy before then. Especially if they will use the string while
the `Address` object is present.