Fix `ucxx::Endpoint::close()` callback function use (#445)
The `callbackFunction`/`callbackData` arguments that can be passed to `ucxx::Endpoint::close()` are turned into a lambda that is itself made into a callback to be executed asynchronously when the close request completes. The objects were being captured as references which is an invalid use due to the local scope of those variables and need to be captured by value since they're `std::function`/`std::shared_ptr`, respectively.

This resolves a bug we previously observed in listener tests that was worked around in a way that for some reason wouldn't hit the issue but was unclear why. That workaround can now be removed.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/ucxx/pull/445