Make send/put buffer pointers const (#516)
To prevent unintentional modification of send/put buffers and also to make it clear for users that those methods are not supposed modify data/pointers, mark them `const`.

In `tagMultiSend` only the data can be `const`, the pointers themselves need to be modifiable by `std::vector`. Perhaps it would make sense to have a specialized container for `tagMultiSend` in the future.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/ucxx/pull/516